renamed to CpuCounter
