+++ src/HealthChecks/IetfResponseWriter.cs
@@ -91,7 +91,7 @@ namespace Ngsa.Application
                 new Dictionary<string, HealthReportEntry> { { CosmosHealthCheck.ServiceId, new HealthReportEntry(res.Status, res.Description, totalTime, res.Exception, res.Data) } },
                 totalTime);
 
-            httpContext.Response.Headers.Add(PerfCounters.CapacityHeader, $"current={PerfCounters.CpuPercent}, target={App.Config.BurstTarget}, max={App.Config.BurstMax}");
+            httpContext.Response.Headers.Add(CpuCounter.CapacityHeader, $"current={CpuCounter.CpuPercent}, target={App.Config.BurstTarget}, max={App.Config.BurstMax}");
 
             // call the response writer
             return IetfResponseWriter(httpContext, rpt);
